Swap 'special' and 'unknown' in enum, so that 'unknown' is default
If 'special' is the default for unknown options then dhcptest refuses
to accept them, with 'Fatal error: Can't specify a value for special
option NN' when there is no format specifed, or with an assertion at
line 616 for basically the same reason, if there is a format specifier.

Swapping so that 'unknown' is the default means that unknown options
can be successfully manually specified.